removed dependencies to AspNetCore and EntityFramework from main project
