+++ src/AutoQueryable.AspNetCore.Filter/AutoQueryable.AspNetCore.Filter.csproj
@@ -34,4 +34,13 @@
     <ProjectReference Include="..\AutoQueryable\AutoQueryable.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Reference Include="Microsoft.EntityFrameworkCore">
+      <HintPath>C:\Users\NilsG.bijtjes\.nuget\packages\microsoft.entityframeworkcore\1.1.2\lib\netstandard1.3\Microsoft.EntityFrameworkCore.dll</HintPath>
+    </Reference>
+    <Reference Include="System.Interactive.Async">
+      <HintPath>C:\Program Files\dotnet\sdk\NuGetFallbackFolder\system.interactive.async\3.1.1\lib\netstandard1.3\System.Interactive.Async.dll</HintPath>
+    </Reference>
+  </ItemGroup>
+
 </Project>
