+++ src/AutoQueryable/Extensions/QueryableExtension.cs
@@ -10,7 +10,6 @@ using AutoQueryable.Core.Aliases;
 using AutoQueryable.Core.Enums;
 using AutoQueryable.Core.Models;
 using AutoQueryable.Models;
-using Microsoft.EntityFrameworkCore;
 
 namespace AutoQueryable.Extensions
 {
@@ -41,7 +40,7 @@ namespace AutoQueryable.Extensions
             }
             if(context.ClauseValueManager.First)
             {
-                return query.FirstOrDefaultAsync();
+                return query.FirstOrDefault();
             }
             var result = query.ToList();
 
@@ -53,23 +52,6 @@ namespace AutoQueryable.Extensions
             return new PagedResult<TEntity>
             {
                 TotalCount = context.TotalCountQuery?.Count() ?? result.Count,
-                RowCount = result.Count,
-                Result = result
-            };
-        }
-        // TODO: Implement WrapWith
-        public static async Task<dynamic> ToAutoQueryListResultAsync<TEntity>(this IQueryable<TEntity> query, IAutoQueryableContext context) where TEntity : class
-        {
-            var result = await query.ToListAsyncSafe();
-
-            //if (!context.ClauseValueManager.Page.HasValue || context.ClauseValueManager.Page.Value == 0)
-            //{
-            //    return result;
-            //}
-            
-            return new PagedResult<TEntity>
-            {
-                TotalCount = context.TotalCountQuery != null ? await context.TotalCountQuery.CountAsyncSafe() : result.Count,
                 RowCount = result.Count,
                 Result = result
             };
@@ -120,25 +102,5 @@ namespace AutoQueryable.Extensions
         //    return wrapper;
         //}
     }
-    public static class EfExtensions
-    {
-        public static Task<List<TSource>> ToListAsyncSafe<TSource>(
-            this IQueryable<TSource> source)
-        {
-            if (source == null)
-                throw new ArgumentNullException(nameof(source));
-            if (!(source is IAsyncEnumerable<TSource>))
-                return Task.FromResult(source.ToList());
-            return source.ToListAsync();
-        }
-        public static Task<int> CountAsyncSafe<TSource>(
-            this IQueryable<TSource> source)
-        {
-            if (source == null)
-                throw new ArgumentNullException(nameof(source));
-            if (!(source is IAsyncEnumerable<TSource>))
-                return Task.FromResult(source.Count());
-            return source.CountAsync();
-        }
-    }
+
 }
